moved tree benefits class
